separate activity from layout (group layout)
