+++ assets/src/components/activities/adaptive/AdaptiveAuthoring.tsx
@@ -1,10 +1,35 @@
-import React from 'react';
+import { NotificationContext } from 'apps/delivery/components/NotificationContext';
+import EventEmitter from 'events';
+import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
-import { AdaptiveModelSchema } from './schema';
-import * as ActivityTypes from '../types';
 import { AuthoringElement, AuthoringElementProps } from '../AuthoringElement';
+import * as ActivityTypes from '../types';
+import { AdaptiveModelSchema } from './schema';
 
-const Adaptive = (props: AuthoringElementProps<AdaptiveModelSchema>) => <p>Adaptive</p>;
+const Adaptive = (props: AuthoringElementProps<AdaptiveModelSchema>) => {
+  const [pusher, _setPusher] = useState(new EventEmitter());
+  console.log('adaptive authoring', props);
+  const parts = props.model?.content?.partsLayout || [];
+  const [selectedPart, setSelectedPart] = useState('');
+
+  const handlePartInit = async (payload: any) => {
+    console.log('AUTHOR PART INIT', payload);
+    return { snapshot: {} };
+  };
+
+  const handlePartClick = (payload: any) => {
+    console.log('AUTHOR PART CLICK', payload);
+    if (props.editMode) {
+      setSelectedPart(payload.id);
+    }
+  };
+
+  return parts && parts.length ? (
+    <NotificationContext.Provider value={pusher}>
+      <div>Parts Layout Render</div>
+    </NotificationContext.Provider>
+  ) : null;
+};
 
 export class AdaptiveAuthoring extends AuthoringElement<AdaptiveModelSchema> {
   render(mountPoint: HTMLDivElement, props: AuthoringElementProps<AdaptiveModelSchema>) {
