+++ assets/src/components/activities/adaptive/components/delivery/PartsLayoutRenderer.tsx
