+++ assets/src/components/activities/adaptive/schema.ts
@@ -1,7 +1,6 @@
 import { ActivityModelSchema, Part, PartComponentDefinition, Transformation } from '../types';
 
 export interface AdaptiveModelSchema extends ActivityModelSchema {
-  // eslint-disable-next-line
   content: {
     custom?: Record<string, any>;
     partsLayout?: PartComponentDefinition[];
